chore(basedpyright): follow basedpiright strict typing suggestions

+++ actions.py
@@ -16,18 +16,19 @@ commands = [
 ]
 
 
-async def execute(message):
+async def execute(message: discord.Message):
     """Execute bot actions"""
     if message.content.startswith(commands[0]):
-        await message.channel.send(f"Available commands are: {', '.join(commands)}")
+        _ = await message.channel.send(f"Available commands are: {', '.join(commands)}")
 
     if message.content.startswith(commands[1]):
         messages_to_delete = int(message.content.split()[1]) + 1
-        messages = [
-            message
-            async for message in message.channel.history(limit=messages_to_delete)
-        ]
-        await message.channel.delete_messages(messages)
+
+        if isinstance(message.channel, discord.TextChannel):
+            messages = [
+                msg async for msg in message.channel.history(limit=messages_to_delete)
+            ]
+            await message.channel.delete_messages(messages)
 
     if message.content.startswith(commands[2]):
         embed = discord.Embed(
@@ -47,15 +48,15 @@ async def execute(message):
         Hidden settings in the lineup tab can optimize your defensive participation. Go to lineups, top right corner click the ratchet wheel ⚙️ next to “quick add” and turn “auto team commands”  “auto convert” on.
         This function will automatically re-place your teams on a base once defeated - even if you are offline.""",
         )
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1124766858902523965/1124782265721958420/Screenshot_20230106-114527.png"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
         embed = discord.Embed()
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1124766858902523965/1124782266086858783/IMG_9920.png?width=1307&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
 
     if message.content.startswith(commands[3]):
         embed = discord.Embed(
@@ -74,55 +75,55 @@ async def execute(message):
         use dynamite. When someone attack us the dynamite will be on the left side (4th pic) when we are attacking the dynamite
         will be on the right side (5th pic) with lines no longer than 120 characters.""",
         )
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1124766858902523965/1124786924050194523/6ED167A9-53C0-4ABA-BD28-C8A0C0FF88AE.jpg"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
         embed = discord.Embed()
-        embed.set_image(
+        _ = embed.set_image(
             url="https://cdn.discordapp.com/attachments/1124766858902523965/1124786924394119168/IMG_3272.png"
         )
-        await message.channel.send(embed=embed)
-        embed.set_image(
+        _ = await message.channel.send(embed=embed)
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1124766858902523965/1124786925098770516/IMG_3275.png?width=1193&height=671"
         )
-        await message.channel.send(embed=embed)
-        embed.set_image(
+        _ = await message.channel.send(embed=embed)
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1124766858902523965/1124786924763238500/IMG_3276.png?width=1193&height=671"
         )
-        await message.channel.send(embed=embed)
-        embed.set_image(
+        _ = await message.channel.send(embed=embed)
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1124766858902523965/1124786925446889532/IMG_3274.png?width=1193&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
 
     if message.content.startswith(commands[4]):
         embed = discord.Embed(title="Tier list")
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1116406148287975464/tierlist310523.png?width=714&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
 
     if message.content.startswith(commands[5]):
         embed = discord.Embed(title="Hero fusion cost")
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1115983977409413201/MLA_Stars.png?width=876&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
 
     if message.content.startswith(commands[6]):
         embed = discord.Embed(title="Wheel of time")
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1111759388533866617/unknown.png?width=632&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
 
     if message.content.startswith(commands[7]):
         embed = discord.Embed(title="Soul vessel")
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1111759375581855834/Soul_Vessel_Yes_Its_Final.png?width=1009&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
 
     if message.content.startswith(commands[8]):
         embed = discord.Embed(
@@ -134,23 +135,23 @@ async def execute(message):
         If you get enough coins to buy the aureate run stones and the HJ copy then wonderful but if you don’t prioritize aureate.
         Last picture is the bonus you get of all 3 resources for every soul root level.""",
         )
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1095812698786242580/IMG_3269.png?width=895&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
         embed = discord.Embed()
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1095812699251814500/IMG_2814.png?width=1193&height=671"
         )
-        await message.channel.send(embed=embed)
-        embed.set_image(
+        _ = await message.channel.send(embed=embed)
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1095812699784478730/Screenshot_2022-09-24-04-32-35-51_64779e9ab56c96eef77afe24eee7e2fc.jpg?width=1440&height=648"
         )
-        await message.channel.send(embed=embed)
-        embed.set_image(
+        _ = await message.channel.send(embed=embed)
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1095812700027756615/IMG_2812.png?width=377&height=671"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
 
     if message.content.startswith(commands[9]):
         embed = discord.Embed(
@@ -162,7 +163,7 @@ async def execute(message):
         If your new challenges score is below the number (in this example case 29.411) you do sweep but if it's higher, you keep the new challenge score. You can calculate any boss with any bonus percentage.
         """,
         )
-        embed.set_image(
+        _ = embed.set_image(
             url="https://media.discordapp.net/attachments/1041544918516113508/1093905919269687457/01E24E33-622F-4380-8E42-8C81054D4EAC.jpg"
         )
-        await message.channel.send(embed=embed)
+        _ = await message.channel.send(embed=embed)
